+++ npm-packages/shared-utils/src/types/types.ts
@@ -131,6 +131,9 @@ export type ExtensionConfig = {
   commandsInfoList?: CommandInfo[];
   // Visibility
   visibility: string;
+  // Recommended dimensions for app view in canvas
+  recommendedHeight?: number;
+  recommendedWidth?: number;
 };
 
 export type CommandInfo = {